Add Servers classes - refactored for providers
